entryAndRoute
